Extract function to generate posix portable filenames
In order to use the function elsewhere extract it to common and make it
public.
